+++ lib/Pear/LocalLoop/Controller/Api/Upload.pm
@@ -75,6 +75,9 @@ has error_messages => sub {
     organisation_name => {
       required => { message => 'organisation name is missing', status => 400 },
     },
+    category => {
+      in_resultset => { message => 'Category is invalid', status => 400 },
+    },
     town => {
       required => { message => 'town/city is missing', status => 400 },
     },
@@ -104,6 +107,7 @@ sub post_upload {
 
   #Check a proper purchase time was submitted
   $validation->optional('purchase_time')->is_full_iso_datetime;
+  $validation->optional('category')->in_resultset( 'id', $c->schema->resultset('Category'));
 
   # First pass of required items
   return $c->api_validation_error if $validation->has_error;
@@ -179,6 +183,7 @@ sub post_upload {
   my $purchase_time = $c->parse_iso_datetime($validation->param('purchase_time') || '');
   $purchase_time ||= DateTime->now();
   my $file = defined $upload ? $c->store_file_from_upload( $upload ) : undef;
+  my $category = $validation->param('category');
   my $distance = $c->get_distance_from_coords( $user->entity->type_object, $organisation );
 
   my $new_transaction = $organisation->entity->create_related(
@@ -203,9 +208,35 @@ sub post_upload {
     );
   }
 
+  if ( defined $category ) {
+    $c->schema->resultset('TransactionCategory')->create({
+      category_id => $category,
+      transaction_id => $new_transaction->id,
+    });
+  }
+
   return $c->render( json => {
     success => Mojo::JSON->true,
     message => 'Upload Successful',
+  });
+}
+
+sub post_category {
+  my $c = shift;
+  my $self = $c;
+
+  my $categories = { ids => [], names => [] };
+
+  my $category_rs = $c->schema->resultset('Category');
+
+  for ( $category_rs->all ) {
+    push @{ $categories->{ ids } }, $_->get_column('id');
+    push @{ $categories->{ names } }, $_->get_column('name');
+  }
+
+  return $self->render( json => {
+    success => Mojo::JSON->true,
+    categories => $categories,
   });
 }