+++ lib/Pear/LocalLoop/Schema/Result/Transaction.pm
@@ -71,4 +71,9 @@ __PACKAGE__->belongs_to(
   { is_deferrable => 0, on_delete => "NO ACTION", on_update => "NO ACTION" },
 );
 
+__PACKAGE__->might_have(
+  "category",
+  "Pear::LocalLoop::Schema::Result::TransactionCategory" => "transaction_id",
+);
+
 1;