+++ share/ddl/SQLite/upgrade/20-21/001-auto.sql
@@ -1,0 +1,13 @@
+-- Convert schema 'share/ddl/_source/deploy/20/001-auto.yml' to 'share/ddl/_source/deploy/21/001-auto.yml':;
+
+;
+BEGIN;
+
+;
+DROP INDEX transaction_category_idx_category_id;
+
+;
+
+;
+
+COMMIT;