+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # Next Release
 
+# v0.10.3
+
+* Added Category and Transaction Category tables to DB
+* Added API for categories in Transactions
+* **Admin Feature** Added ability to add and delete categories
+* **Admin Feature** Added ability to view transaction category
+* Fixed all relevant tests to match
+
 # v0.10.2
 
 * Added fairly traded column for organisations